Remove unnecessary integration tests to regular client tests
