+++ contracts/test-integration/BlsProvider.test.ts
@@ -131,70 +131,6 @@ describe("BlsProvider", () => {
     expect(spy).to.have.been.called.exactly(2);
   });
 
-  it("should throw an error when sending multiple signed operations to sendTransaction", async () => {
-    // Arrange
-    const expectedAmount = parseEther("1");
-    const verySafeFee = parseEther("0.1");
-    const firstRecipient = ethers.Wallet.createRandom().address;
-    const secondRecipient = ethers.Wallet.createRandom().address;
-
-    const firstActionWithSafeFee = blsProvider._addFeePaymentActionWithSafeFee(
-      [
-        {
-          ethValue: expectedAmount,
-          contractAddress: firstRecipient,
-          encodedFunction: "0x",
-        },
-      ],
-      verySafeFee,
-    );
-    const secondActionWithSafeFee = blsProvider._addFeePaymentActionWithSafeFee(
-      [
-        {
-          ethValue: expectedAmount,
-          contractAddress: secondRecipient,
-          encodedFunction: "0x",
-        },
-      ],
-      verySafeFee,
-    );
-
-    const nonce = await blsSigner.wallet.Nonce();
-
-    const firstOperation = {
-      nonce,
-      actions: [...firstActionWithSafeFee],
-    };
-    const secondOperation = {
-      nonce,
-      actions: [...secondActionWithSafeFee],
-    };
-
-    const firstBundle = await blsSigner.wallet.signWithGasEstimate(
-      firstOperation,
-    );
-    const secondBundle = await blsSigner.wallet.signWithGasEstimate(
-      secondOperation,
-    );
-
-    const aggregatedBundle = blsSigner.wallet.blsWalletSigner.aggregate([
-      firstBundle,
-      secondBundle,
-    ]);
-
-    // Act
-    const result = async () =>
-      await blsProvider.sendTransaction(
-        JSON.stringify(bundleToDto(aggregatedBundle)),
-      );
-
-    // Assert
-    await expect(result()).to.rejectedWith(
-      Error,
-      "Can only operate on single operations. Call provider.sendTransactionBatch instead",
-    );
-  });
-
   it("should throw an error when sending a modified signed transaction", async () => {
     // Arrange
     const address = await blsSigner.getAddress();
